SWMedia-Bot/requirements.txt
aiogram==2.7
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv~=0.13.0
requests~=2.23.0